# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GachaPull - Pokemon & One Piece Gacha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CARDS
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List all cards
      parameters:
        - in: query
          name: franchise
          schema:
            type: string
            enum: [pokemon, onepiece, all]
        - in: query
          name: rarity
          schema:
            type: string
            enum: [common, rare, super_rare, ultra_rare, legendary]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Cards list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardListResponse"
    post:
      operationId: createCard
      tags: [cards]
      summary: Create a card (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBody"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
...
  /cards/{cardId}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get card detail
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    put:
      operationId: updateCard
      tags: [cards]
      summary: Update card (admin)
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBody"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: deleteCard
      tags: [cards]
      summary: Delete card (admin)
      parameters:
        - in: path
          name: cardId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PACKS
...
  /gacha/pull:
    post:
      operationId: gachaPull
      tags: [gacha]
      summary: Pull from a pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GachaPullBody"
      responses:
        "200":
          description: Pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GachaPullResult"
        "400":
          description: Insufficient coins or pack unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gacha/history:
    get:
      operationId: getGachaHistory
      tags: [gacha]
      summary: Get my pull history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Pull history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GachaHistoryResponse"

  # COLLECTION
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
    RegisterBody:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
    UserProfile:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        avatarUrl:
          type: string
          nullable: true
        totalSpentUsd:
          type: number
          nullable: true
        coinsBalance:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time

    # Card schemas
    Card:
      type: object
      required: [id, name, franchise, rarity, imageUrl, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        rarity:
          type: string
          enum: [common, rare, super_rare, ultra_rare, legendary]
        imageUrl:
          type: string
        description:
          type: string
          nullable: true
        pullCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    CardListResponse:
      type: object
      required: [cards, total, page, limit]
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateCardBody:
      type: object
      required: [name, franchise, rarity, imageUrl]
      properties:
        name:
          type: string
        franchise:
          type: string
          enum: [pokemon, onepiece]
        rarity:
          type: string
          enum: [common, rare, super_rare, ultra_rare, legendary]
        imageUrl:
          type: string
        description:
          type: string
          nullable: true

    # Pack schemas
...
    GachaPullBody:
      type: object
      required: [packId, pullCount]
      properties:
        packId:
          type: integer
        pullCount:
          type: integer
          enum: [1, 10]
          description: Pull 1 or 10 cards
    GachaPullResult:
      type: object
      required: [cards, coinsSpent, coinsRemaining]
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/PulledCard"
        coinsSpent:
          type: integer
        coinsRemaining:
          type: integer
    PulledCard:
      type: object
      required: [card, isNew]
      properties:
        card:
          $ref: "#/components/schemas/Card"
        isNew:
          type: boolean
          description: True if this is the first time user got this card
    GachaHistoryResponse:
      type: object
      required: [pulls, total, page, limit]
      properties:
        pulls:
          type: array
          items:
            $ref: "#/components/schemas/GachaPullRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    GachaPullRecord:
      type: object
      required: [id, pack, card, pulledAt]
      properties:
        id:
          type: integer
        pack:
          type: object
          required: [id, name]
          properties:
            id:
              type: integer
            name:
              type: string
        card:
          $ref: "#/components/schemas/Card"
        pulledAt:
          type: string
          format: date-time

    # Collection schemas
...